refactor LLMClient
